services/api/requirements.txt
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-cors-headers==4.4.0
django-filter==24.2
django-redis==5.0.3
django-celery-beat==2.6.0
celery==5.4.0
redis==5.0.7
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==22.0.0
uvicorn[standard]==0.30.1
pytest==8.2.2
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0
freezegun==1.5.1
requests==2.32.3
responses==0.25.3
cryptography==42.0.8
google-api-python-client==2.134.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
stripe==9.10.0
clickhouse-driver==0.2.7